requirements.txt
bert-serving-client
python-dotenv
mongoengine
typesense
celery==4.2.0rc4
redis
elasticsearch
